#PROG-15
Continuação implementação classe cliente
