+++ Testes/nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=C:\\Users\\ariad\\AppData\\Roaming\\NetBeans\\16\\build.properties
