+++ src/main/java/br/com/vitavault/MainAri.java
@@ -6,7 +6,6 @@ package br.com.vitavault;
 
 import br.com.vitavault.exceptions.GerenciadorClientesException;
 import br.com.vitavault.model.Cliente;
-import br.com.vitavault.model.GerenciadorClientes;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -16,13 +15,6 @@ import java.util.List;
  */
 public class MainAri {
     public static void main(String[] args) throws GerenciadorClientesException {
-        List<Cliente> clientes = new ArrayList();
-//        Cliente oCliente = new Cliente("Teste", "03921", "Rua teste", 212212);
-//        Cliente oCliente1 = new Cliente("Teste", "03", "Rua teste", 212212);
-//        
-//        GerenciadorClientes gerenciarClientes = new GerenciadorClientes(clientes);
-//        gerenciarClientes.cadastrarCliente(oCliente);
-//        gerenciarClientes.removerCliente(oCliente);
-        
+//         TelaLogin tela = new TelaLogin();
     }
 }
